feat(KMM): add Claude/Anthropic support and fix iOS build for modern SDKs
- Add AIClient interface and AnthropicChatClient for multi-provider support
- Add OpenAIChatClient wrapping existing openai-client dependency
- Update Config/Runner to support AIProvider enum (OpenAI, Anthropic)
- Fix cinterop for iOS 26+ SDKs: replace module-based XCTest import with
  minimal xctest_stub.h using Foundation-only headers and -fno-modules removed
- Fix ObjC protocol naming: stub protocols named without "Protocol" suffix so
  Kotlin/Native generates the expected XCUIElementSnapshotProtocol names
- Add XCUIAutomation to linker opts (re-exported from XCTest, explicit for clarity)
- Upgrade Gradle wrapper 7.5 → 8.7 and KMM plugin 1.8.x → 2.1.20
- Add CLAUDE.md with project structure and build instructions

+++ sdk/testpilot/src/iosMain/kotlin/co/work/testpilot/TestActorIOS.kt
@@ -39,7 +39,7 @@ class TestActorIOS(val testCase: XCTestCase) : TestActor<AppUISnapshotIOS, Testa
             is Instruction.ScrollUp -> xcApp.swipeDownWithVelocity(XCUIGestureVelocitySlow)
             is Instruction.ScrollDown -> xcApp.swipeUpWithVelocity(XCUIGestureVelocitySlow)
             is Instruction.GoBack -> {
-                val match = xcApp.navigationBars.buttons.elementBoundByIndex(0)
+                val match = xcApp.navigationBars.buttons.elementBoundByIndex(0.toULong())
                 match.tap()
             }
         }
